ExpireOnStartup
Configuration option to run expiry commands on startup (For those
players that don't sign on the day their package expires.) Instead, it
will run when the server starts up.

There is still an issue with it not expiring if the date has already
passed. I'm unfamiliar with Java dates. Could use a Dev to help out with
this. I'll probably add a command to check expire dates  / automatically
expire packages through commands for those that the checkers miss.