Bug: Player stays in HashMap if frame is linked.
If the player clicked an item frame that was already linked, it kept
them in the hashmap, meaning they had to link the package somewhere.
This has been fixed so it removes the player from the HashMap in the
event the player can't link a package to a certain frame.